% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calib.R
\name{calib}
\alias{calib}
\title{The function calib}
\usage{
calib(conc, OD)
}
\arguments{
\item{conc}{is a numeric vector with cocnentrations}

\item{OD}{is a vector with ODs}

\item{...}{currently disregarded}
}
\value{
the coefficients of the calibration model
}
\description{
Gives coefficients for calcibration curve
}
\details{
Some more details here
}
\examples{
conc <-1:10
OD<-1:10 + rnorm(10)
calib(conc,OD)
}
